메인로직수정
